Moved API code to API project.
