+++ OnlineShop.API/Controllers/BaseAPIController.cs
@@ -5,6 +5,11 @@ namespace OnlineShop.API.Controllers
 {
     public abstract class BaseAPIController : ApiController
     {
+        protected IOrderService _orderService;
+        protected BaseAPIController(IOrderService orderService)
+        {
+            _orderService = orderService;
+        }
 
     }
 }
