+++ OnlineShop.API/Models/OrderItem.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 
-namespace OnlineShop.Models
+namespace OnlineShop.API.Models
 {
     public class OrderItem
     {
