+++ OnlineShop.API/Models/User.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Web;
 
-namespace OnlineShop.Models
+namespace OnlineShop.API.Models
 {
     public class User
     {
