+++ OnlineShop/Controllers/HomeController.cs
@@ -1,9 +1,4 @@
-﻿using OnlineShop.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using System.Web.Mvc;
+﻿using System.Web.Mvc;
 
 namespace OnlineShop.Controllers
 {
