+++ OnlineShop/Global.asax.cs
@@ -2,7 +2,6 @@
 using System.Web;
 using System.Web.Mvc;
 using System.Web.Routing;
-using System.Web.Http;
 using System.Web.Optimization;
 
 namespace OnlineShop
@@ -13,7 +12,6 @@ namespace OnlineShop
         {
             // Code that runs on application startup
             AreaRegistration.RegisterAllAreas();
-            GlobalConfiguration.Configure(WebApiConfig.Register);
             RouteConfig.RegisterRoutes(RouteTable.Routes);            
             BundleConfig.RegisterBundles(BundleTable.Bundles);
         }
